% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_centers.R
\name{add_centers}
\alias{add_centers}
\title{Add centroids manually}
\usage{
add_centers(centers, xmap, x = NULL, y = NULL, phase = if
  (is.null(x)) NULL else paste0("P", seq_along(x)), i = data.frame(x, y,
  phase), ..., saveas = "center_add.csv")
}
\arguments{
\item{centers}{c-by-p matrix returned by \code{find_centers()} or by manually; c clusters and p features. Used to guess initial centers (or centroids) of clusters. A value returned by , typically \code{data.frame} or \code{matrix}, indicating initial guess centers (or centroids) or clusters. See \code{find_centers()}.}

\item{xmap}{an \code{qm_xmap} class object generated by `read_xmap()``}

\item{x}{
A n-by-p training data matrix; n observations and p features. Used to train the
classifier.
}

\item{y}{Vectors of \code{integer} specifying x- and y-coordinates of pixels of mapping data.
In other words, x- and y- corresponds to n-th column and row of a matrix.
Defaults are set to \code{NULL} to control}

\item{phase}{Name of phases comprising the pixels to be picked.
If not manually supplied \code{phase} is supplied by \code{c('P1', 'P2', ...)}.}

\item{i}{A data frame with clumns \code{x}, \code{y}, and \code{phase}.
\code{i} have priority to \code{x}, \code{y}, and \code{phase}.`}

\item{...}{ignored}

\item{saveas}{File name to save result. FALSE if not saving.}
}
\description{
Add centroids manually by picking X-ray counts of selected pixels in X-ray maps.
}
\examples{
centers <- data.frame(
  phase = 'a',
  A = 1,
  B = 1
)
xmap <- list(
  A = data.frame(1:2),
  B = data.frame(1:2)
)
class(xmap) <- c('qm_xmap', 'list')
add_centers(centers, xmap, x = 1, y = 2, saveas = FALSE)

}
\seealso{
\code{\link[=pick]{pick()}}

\code{\link[=find_centers]{find_centers()}}
}
