% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-autoplot-qntmap.R
\name{autoplot-qntmap}
\alias{autoplot-qntmap}
\alias{autoplot.qntmap}
\alias{autoplot.qm_xmap}
\alias{autoplot.qm_cluster}
\title{Create a ggplot object for objects created by qntmap package}
\usage{
\method{autoplot}{qntmap}(
  object,
  zname = setdiff(names(object), c("x", "y"))[[1L]],
  zlim = NULL,
  colors = c("magma", "viridis", "gray"),
  unit = c("px", "um", "mm", "cm"),
  ...
)

\method{autoplot}{qm_xmap}(
  object,
  zname = setdiff(names(object), c("x", "y"))[[1L]],
  zlim = NULL,
  colors = c("magma", "viridis", "gray"),
  unit = c("px", "um", "mm", "cm"),
  ...
)

\method{autoplot}{qm_cluster}(
  object,
  zname = "cluster",
  zlim = NULL,
  colors = c("magma", "viridis", "gray"),
  unit = c("px", "um", "mm", "cm"),
  ...
)
}
\arguments{
\item{object}{an object, whose class will determine the behaviour of autoplot}

\item{zname}{Name of z-axis (fill).
If unset, first column except "x" and "y" is chosen for objects with
\code{qm_xmap} or \code{qntmap} classes, and "cluster" is chosen for the \code{qm_cluster}
class object.}

\item{zlim}{Range limit of z-coordinates. This is neglected if z is discrete.}

\item{colors}{One of "magma", "viridis", or "gray". Applicable when z-coordinates is numeric.}

\item{unit}{Unit of x- and y-axis ("px", "um", "mm" or "cm").}

\item{...}{other arguments passed to specific methods}
}
\description{
Create a ggplot object for objects created by qntmap package
}
