% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cipois.R
\name{cipois}
\alias{cipois}
\title{confidence interval of observed data from poisson process}
\usage{
cipois(x, vars = names(x), offset = 1, conf.level = 0.95)
}
\arguments{
\item{x}{data frame containing variable whose data is from poisson process}

\item{vars}{a character vector indicating which variables of x are poisson process data}

\item{offset}{vector, list, or data frame indicating rate of poisson process data.}

\item{conf.level}{confidence level for the returned confidence interval.}
}
\description{
confidence interval of observed data from poisson process
}
