% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_xmap}
\alias{cluster_xmap}
\title{Cluster mapping data into mineral species}
\usage{
cluster_xmap(xmap, centers, elements = intersect(names(xmap),
  colnames(centers)), saving = TRUE, group_cluster = FALSE)
}
\arguments{
\item{xmap}{a \code{qm_xmap} class object returned by \code{read_xmap}}

\item{centers}{c-by-p matrix returned by \code{find_centers()} or by manually; c clusters and p features. Used to guess initial centers (or centroids) of clusters. A value returned by , typically \code{data.frame} or \code{matrix}, indicating initial guess centers (or centroids) or clusters. See \code{find_centers()}.}

\item{elements}{A character vector indicating which elements to be utilized in cluster analysis. \code{NULL}, in default, selects as much elements as possible are utilized in cluster analysis.}

\item{saving}{\code{TRUE} or \code{FALSE} to save result.}

\item{group_cluster}{\code{TRUE} (default) or \code{FALSE} to integrate same phase subgrouped using suffix. For example, when there are clusters named as Pl_NaRich and Pl_NaPoor, they are grouped as Pl.}
}
\description{
Cluster mapping data into mineral species
}
