% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{defunct}
\alias{defunct}
\alias{qntmap_quantify}
\alias{qltmap_cls_centers}
\alias{qltmap_load}
\alias{qntmap_cls_pois}
\alias{qnt_load}
\title{Defunct functions}
\usage{
qntmap_quantify(...)

qltmap_cls_centers(...)

qltmap_load(...)

qntmap_cls_pois(...)

qnt_load(...)
}
\arguments{
\item{...}{Ignored}
}
\description{
Defunct functions
}
