% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epma_tidy.R
\name{epma_tidy}
\alias{epma_tidy}
\title{compile epma data}
\usage{
epma_tidy(wd = NULL, dir_map, phase_list = NULL, RDS_cluster = NULL,
  qnt = qnt_load(wd, phase_list = phase_list),
  qltmap = qltmap_load(dir_map), cluster = if (is.null(RDS_cluster)) NA else
  readRDS(RDS_cluster))
}
\arguments{
\item{wd}{path to the directory containing .qnt directory}

\item{dir_map}{path to the directory containing mapping analysis e.g., ./.map/1/}

\item{phase_list}{path to the csv file containing columns indicating phase of each analysis and true or false to use it for quantifying.}

\item{RDS_cluster}{path to the clustering result}

\item{qnt}{object of class qnt}

\item{qltmap}{object of class qltmap}

\item{cluster}{object of class PoiClaClu}
}
\description{
compile epma data
}
