% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centers.R
\name{find_centers}
\alias{find_centers}
\title{Initialize centroids for \code{\link[=cluster_xmap]{cluster_xmap()}}}
\usage{
find_centers(xmap, qnt, fine_phase = NULL, saveas = "centers0.csv")
}
\arguments{
\item{xmap}{an \code{qm_xmap} class object generated by \code{\link[=read_xmap]{read_xmap()}}}

\item{qnt}{an \code{qm_qnt} class object generated by \code{\link[=read_qnt]{read_qnt()}}`}

\item{fine_phase}{A character vector to specify fine grained phases
which tend to comprise multi-phase pixels e.g., c('Pl', 'Amp')}

\item{saveas}{File name to save result. FALSE if not saving.}
}
\description{
Initialize centroids for \code{\link[=cluster_xmap]{cluster_xmap()}}
}
