% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centers_kpp.R
\name{find_centers_kpp}
\alias{find_centers_kpp}
\title{Generate initial centroids for clustering randomly}
\usage{
find_centers_kpp(x, k, given = NULL)
}
\arguments{
\item{x}{An object which can be coerced to \code{matrix}, typically \code{matrix} itself or \code{data.frame}.}

\item{k}{A number of clusters.}

\item{given}{Given centers. Not yet implemented.}
}
\description{
Generate initial centroids for clustering randomly
}
