% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{flag0}
\alias{flag0}
\title{return integer as character flagged with 0}
\usage{
flag0(...)
}
\arguments{
\item{...}{list of integer vectors of same length}
}
\description{
return integer as character flagged with 0
}
