% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group-cluster.R
\name{group_cluster}
\alias{group_cluster}
\title{Group clusters who share same prefix}
\usage{
group_cluster(x, saving = TRUE, suffix = "_.*")
}
\arguments{
\item{x}{\code{qm_cluster} class object returned by \code{\link[=cluster_xmap]{cluster_xmap()}}.}

\item{saving}{\code{TRUE} or \code{FALSE} to save result (default: \code{TRUE}).}

\item{suffix}{A regular expression of suffix (default: \code{'_.*'})}
}
\description{
When data points are assigned to clusters A_1 and A_2,
their clusters are renamed to be A by matching regular expressions.
}
