% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{group_subclusters}
\alias{group_subclusters}
\title{Group sub-clusters who share same prefix}
\usage{
group_subclusters(x, suffix = "_.*")
}
\arguments{
\item{x}{The \code{qm_cluster} class object.}

\item{suffix}{A regular expression of suffix of cluster names.
Clusters with the same prefix comprise a super cluster.
For example, "Pl_NaRich" and "Pl_NaPoor" becomes "Pl" cluster if
\code{suffix = "_.*"} (default).}
}
\description{
When data points are assigned to clusters A_1 and A_2,
their clusters are renamed to be A by matching regular expressions.
}
