% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{hmean}
\alias{hmean}
\alias{vmean}
\title{Calculate means for horizontal and vertical directions}
\usage{
hmean(x, step = attr(x, "step")[1L], ...)

vmean(x, step = attr(x, "step")[1L], ...)
}
\arguments{
\item{x}{Mapping data (currently restricted to an object of class "qntmap").}

\item{step}{A step size as a numeric value (i.e., length of pixel's sides).
Specify when \code{x} is produced by \code{qntmap} before 0.3.4.}

\item{...}{Arguments passed to \code{mean.qntmap()}
if x is a \code{qntmap} class object returned by \code{quantify()} or \code{qntmap()}.
Otherwise ignored.}
}
\description{
Calculate means for horizontal and vertical directions for mapping data.
}
\seealso{
mean.qntmap
}
