% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-.R
\name{mean}
\alias{mean}
\alias{mean.qntmap}
\alias{mean.qm_xmap}
\alias{mean.qm_cluster}
\title{Arithmetric mean for objects generated by \code{qntmap} package}
\usage{
\method{mean}{qntmap}(x, index = "Whole area", cluster = NULL,
  density = NULL, ...)

\method{mean}{qm_xmap}(x, index = "Whole area", cluster = NULL,
  density = NULL, ...)

\method{mean}{qm_cluster}(x, index = "Whole area", cluster = NULL,
  density = NULL, ...)
}
\arguments{
\item{x}{A \code{qntmap} class object returned by \code{\link[=quantify]{quantify()}} or \code{\link[=qntmap]{qntmap()}}.}

\item{index}{A vector of length `1`` or length equal to number of pixels in map.
An index can be created using mask image through \code{\link[=segment]{segment()}}.}

\item{cluster}{A result of \code{cluster_xmap()} (i.e., \code{qm_cluster} class object).
\code{cluster} is required when specifying \code{density}.}

\item{density}{A named numeric vector specifying density of all the phases in \code{x}.
(e.g., \code{c(Qtz = 2.6, Ab = 2.6)}).}

\item{...}{Ignored}
}
\description{
S3 methods for the arithmetric mean.
See \code{\link[base:mean]{base::mean()}}  for general use of \code{mean()}.
}
\section{mean.qntmap}{

A returning value is a \code{\link{data.frame}} whose first column lists elements,
and second to last columns lists \code{mean} values of each elements by index.
}

\seealso{
\code{\link[base:mean]{base::mean()}}, \code{\link[=segment]{segment()}}, \code{\link[=quantify]{quantify()}}, \code{\link[=qntmap]{qntmap()}}
}
