% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{mycolors}
\alias{mycolors}
\title{color LUT}
\usage{
mycolors(palette = c("pcol", "gray"), n = NULL, dec = FALSE)
}
\arguments{
\item{palette}{palette to be used as LUT. either pcol (= pseudocolor) or gray. It is ignored when LUT is specified.}

\item{n}{number of output colors. When n is more than number of colors in the specified palette, output contains duplicated colors.}

\item{dec}{FALSE in default outputs a vector of RGB colors. TRUE outputs matrix whose columns are R, G, and B, and whose values are in decimals.}
}
\description{
color LUT
}
