% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick.R
\name{pick}
\alias{pick}
\title{Pick values from mapping data}
\usage{
pick(
  .df,
  x = NULL,
  y = NULL,
  phase = if (is.null(x)) NULL else paste0("P", seq_along(x)),
  i = data.frame(x, y, phase),
  ...
)
}
\arguments{
\item{.df}{Mapping data. Currentl version only supports \code{qm_xmap} object.}

\item{x, y}{Integer vectors specifying x- and y-coordinates of pixels of mapping data.
In other words, x- and y- corresponds to n-th column and row of a matrix.}

\item{phase}{Name of phases comprising the pixels to be picked.
If not manually supplied \code{phase} is supplied by \code{c('P1', 'P2', ...)}.}

\item{i}{A data frame with clumns \code{x}, \code{y}, and \code{phase}.
\code{i} have priority to \code{x}, \code{y}, and \code{phase}.`}

\item{...}{Ignored}
}
\value{
\code{data.frame}
}
\description{
Pick values from mapping data based on coordinates of x and y.
}
\seealso{
\code{\link[=add_centers]{add_centers()}}
}
