% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-.R
\name{plot-qntmap}
\alias{plot-qntmap}
\alias{plot.qntmap}
\alias{plot.qm_xmap}
\alias{plot.qm_cluster}
\title{Plot methods for \code{qntmap} package}
\usage{
\method{plot}{qntmap}(x, y = setdiff(names(x), c("x", "y"))[1L],
  zname = y, zlim = NULL, colors = c("magma", "viridis", "gray"),
  interactive = TRUE, ...)

\method{plot}{qm_xmap}(x, y = setdiff(names(x), c("x", "y"))[1L],
  zname = y, zlim = NULL, colors = c("magma", "viridis", "gray"),
  interactive = TRUE, ...)

\method{plot}{qm_cluster}(x, y = "cluster", zname = y, zlim = NULL,
  colors = c("magma", "viridis", "gray"), interactive = TRUE, ...)
}
\arguments{
\item{x}{An object of class \code{qntmap}, \code{qm_cluster}, or \code{qm_xmap},
returned by \code{\link[=quantify]{quantify()}}, \code{\link[=qntmap]{qntmap()}}, and \code{\link[=read_xmap]{read_xmap()}}, respectively.}

\item{zname, y}{A string specifying a component of \code{x} to determine colors to fill the map.
\code{y} is the alias of \code{zname}.}

\item{zlim}{A range of z.}

\item{colors}{A color scale "viridis" (default) or "gray" applicable when fill is continuous.}

\item{interactive}{\code{TRUE} (default) produces plots with shiny WebUI, and
\code{FALSE} produces plots with \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{...}{Arguments passed to internal functions.}
}
\description{
S3 methods to plot object with original classes in \code{qntmap} package.
See \code{\link[graphics:plot]{graphics::plot()}} for general use of \code{plot}.
Mapping data (\code{qm_xmap} and \code{qm_qntmap} classes) are visualized by heat maps.
}
\examples{
# qm_raster class object
d <- data.frame(expand.grid(x = 1:5, y = 1:5), fill = runif(5))
class(d) <- c("qm_raster", class(d))
plot(d, "fill", interactive = FALSE)
}
\seealso{
\code{\link[graphics:plot]{graphics::plot()}}
}
