% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot.qm_xmap}
\alias{plot.qntmap}
\title{Plot methods for \code{qntmap} package}
\usage{
\method{plot}{qm_xmap}(x, y = setdiff(names(x), c("x", "y"))[1],
  legend_fill = y, interactive = TRUE, ..., shiny = FALSE)

\method{plot}{qntmap}(x, y = setdiff(names(x), c("x", "y"))[1],
  legend_fill = y, interactive = TRUE, ..., shiny = FALSE)
}
\arguments{
\item{x}{\code{qm_xmap} or \code{qntmap} class objects returned by
\code{\link[=read_xmap]{read_xmap()}}, \code{\link[=quantify]{quantify()}}, or \code{\link[=qntmap]{qntmap()}}.}

\item{y}{A string specifying a component of \code{x} to determine colors to fill the map.}

\item{legend_fill}{A string to specify legend name for fill.
Default value is taken from \code{y}.}

\item{interactive}{\code{TRUE} returns \code{plotly} object (default), and
\code{FALSE} returns \code{ggplot} object.}

\item{...}{ignored}

\item{shiny}{See plots using Shiny (default: \code{FALSE})}
}
\description{
S3 methods to plot object with original classes in \code{qntmap} package.
See \code{\link[graphics:plot]{graphics::plot()}} for general use of \code{plot}.
Mapping data (\code{qm_xmap} and \code{qm_qntmap} classes) are visualized by heat maps.
}
\examples{
# qm_xmap class object
xm <- list(A = as.data.frame(matrix(runif(25), 5)))
class(xm) <- c('qm_xmap', 'list')
plot(xm)

# qntmap class object
qm <- list(A = list(wt = as.data.frame(matrix(runif(25), 5))))
class(qm) <- c( 'qntmap', 'list')
plot(qm)

}
\seealso{
\code{\link[graphics:plot]{graphics::plot()}}
}
