% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_centers.R
\name{qltmap_cls_centers}
\alias{qltmap_cls_centers}
\title{Generate initial centroids for clustering}
\usage{
qltmap_cls_centers(qnt = NULL, qltmap = NULL, cnd_qltmap = "0.cnd",
  wd = NULL, saving = "centers_initial0.csv")
}
\arguments{
\item{qnt}{path to the .RDS file which compiles quantitative data}

\item{qltmap}{path to the .RDS file which compiles mapping data}

\item{cnd_qltmap}{path to the condition file of mapping data}

\item{wd}{working directory which contains mapping data}

\item{saving}{file name to save. FALSE if not saving.}
}
\description{
Generate initial centroids for clustering
}
