% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_centers_random.R
\name{qltmap_cls_centers_random}
\alias{qltmap_cls_centers_random}
\title{Generate initial centroids for clustering randomly}
\usage{
qltmap_cls_centers_random(k, given = NULL, qltmap = NULL, wd = NULL)
}
\arguments{
\item{k}{number of clusters}

\item{given}{given centers}

\item{qltmap}{path to the .RDS file which compiles mapping data}

\item{wd}{working directory which contains mapping data}
}
\description{
Generate initial centroids for clustering randomly
}
