% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_centers_random2.R
\name{qltmap_cls_centers_random2}
\alias{qltmap_cls_centers_random2}
\title{Generate initial centroids for clustering randomly}
\usage{
qltmap_cls_centers_random2(k = 2, centers_manual = NULL, qltmap = NULL,
  wd = NULL)
}
\arguments{
\item{k}{number of clusters}

\item{centers_manual}{partial and manual selection of initial centers}

\item{qltmap}{path to the .RDS file which compiles mapping data}

\item{wd}{working directory which contains mapping data}
}
\description{
Generate initial centroids for clustering randomly
}
