% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_pois_integrate.R
\name{qltmap_cls_pois_integrate}
\alias{qltmap_cls_pois_integrate}
\title{integrate clusters named with same prefix (e.g., A_1 and A_2 are integrated to be A)}
\usage{
qltmap_cls_pois_integrate(result, wd = NULL, saving = TRUE)
}
\arguments{
\item{result}{result of qltmap_cls_pois}

\item{wd}{Directory containing mapping data. If NULL, current directory is wd.}

\item{saving}{TRUE or FALSE to save result}
}
\description{
integrate clusters named with same prefix (e.g., A_1 and A_2 are integrated to be A)
}
