% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_save.R
\name{qltmap_cls_save}
\alias{qltmap_cls_save}
\title{save the result of qltmap_cls_pois() or qltmap_cls_pois_integrate() as RDS file and png file}
\usage{
qltmap_cls_save(result, name_write, components = NULL)
}
\arguments{
\item{result}{result of clustering}

\item{name_write}{name of result to be written}

\item{components}{components in RDS}
}
\description{
save the result of qltmap_cls_pois() or qltmap_cls_pois_integrate() as RDS file and png file
}
