% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmap.R
\name{qltmap_load}
\alias{qltmap_load}
\title{(DEPRECATED) Use read_xmap}
\usage{
qltmap_load(wd = ".", DT = 0, RDS, renew = FALSE, saving = TRUE)
}
\arguments{
\item{wd}{directory path containing mapping data (e.g., ./.map/.1)}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{RDS}{ignored.}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}
}
\description{
(DEPRECATED) Use read_xmap
}
