% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap-package.R
\docType{package}
\name{qntmap-package}
\alias{qntmap-package}
\title{Enhance quantitative analysis of EPMA maps with QntMap}
\description{
QntMap enables conversion of EPMA X-ray maps to mass concetration maps.
See website for more information \url{https://qntmap.atusy.net}.
}
\references{
Yasumoto, A., Yoshida, K., Kuwatani, T., Nakamura, D., Svojtka, M., &
Hirajima, T. (2018).
A rapid and precise quantitative electron probe chemical mapping technique
and its application to an ultrahigh-pressure eclogite
from the Moldanubian Zone of the Bohemian Massif
(Nové Dvory, Czech Republic).
American Mineralogist, 103(10), 1690-1698,
\url{https://doi.org/10.2138/am-2018-6323CCBY}.
}
