% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap_AB.R
\name{qntmap_AB_fix}
\alias{qntmap_AB_fix}
\title{fix AB value when composition of certain phases are constant}
\usage{
qntmap_AB_fix(AB, fix = NULL, X, fine_th = 0.9, qltmap)
}
\arguments{
\item{AB}{AB}

\item{fix}{csv file indicating composition of the phases. NULL returns input AB}

\item{X}{membership}

\item{fine_th}{threshold of X}
}
\description{
fix AB value when composition of certain phases are constant
}
