% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{qntmap_cls_pois}
\alias{qntmap_cls_pois}
\title{(Deprecated) Use \code{cluster_xmap()}}
\usage{
qntmap_cls_pois(centers_initial, qltmap = NULL, elements = NULL,
  wd = ".", saving = TRUE, integration = TRUE)
}
\arguments{
\item{centers_initial}{Equivalent to \code{centers_initial} parameter of \code{cluster_xmap}}

\item{qltmap}{Equivalent to \code{xmap} parameter of \code{cluster_xmap}}

\item{elements}{A character vector indicating which elements to be utilized in cluster analysis. \code{NULL}, in default, selects as much elements as possible are utilized in cluster analysis.}

\item{wd}{Path to the working directory.}

\item{saving}{\code{TRUE} or \code{FALSE} to save result.}

\item{integration}{Equivalent to \code{group_cluster} parameter of \code{cluster_xmap}}
}
\description{
(Deprecated) Use \code{cluster_xmap()}
}
