% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap_bad.R, R/qntmap_quantify.R
\name{qntmap_quantify}
\alias{qntmap_quantify}
\alias{qntmap_quantify}
\title{quantify qualtitative mapping data}
\usage{
qntmap_quantify(wd = NULL, phase_list = NULL, dir_map, RDS_cluster,
  maps_x = NULL, maps_y = NULL, fine_phase = NULL, fine_th = 0.9,
  qnt = qnt_load(wd, phase_list = phase_list),
  qltmap = qltmap_load(dir_map), cluster = readRDS(RDS_cluster))

qntmap_quantify(wd = NULL, phase_list = NULL, dir_map, RDS_cluster,
  maps_x = NULL, maps_y = NULL, fine_phase = NULL, fine_th = 0.9,
  qnt = qnt_load(wd, phase_list = phase_list),
  qltmap = qltmap_load(dir_map), cluster = readRDS(RDS_cluster))
}
\arguments{
\item{wd}{working directory which contains .qnt and .map directories}

\item{phase_list}{aaa}

\item{dir_map}{directory containing map data to be quantified}

\item{RDS_cluster}{path to the output RDS file of clustering. NULL in default look for the newest one in dir_map/clustering}

\item{maps_x}{aaa}

\item{maps_y}{aaa}

\item{fine_phase}{aaa}

\item{fine_th}{0.9}

\item{qnt}{object of class qnt}

\item{qltmap}{object of class qltmap}

\item{cluster}{object of class PoiClaClu}

\item{path_phase}{aaa}

\item{wd}{working directory which contains .qnt and .map directories}

\item{dir_map}{directory containing map data to be quantified}

\item{maps_x}{aaa}

\item{maps_y}{aaa}

\item{RDS_cluster}{path to the output RDS file of clustering. NULL in default look for the newest one in dir_map/clustering}

\item{fine_phase}{aaa}

\item{fine_th}{0.9}
}
\description{
quantify qualtitative mapping data

quantify qualtitative mapping data
}
