% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantify X-ray maps}
\usage{
quantify(xmap, qnt, cluster, maps_x = attr(xmap, "pixel")[1],
  maps_y = attr(xmap, "pixel")[2], fine_phase = NULL, fine_th = 0.9,
  fixAB = NULL, fixB = NULL, saving = TRUE)
}
\arguments{
\item{xmap}{\code{qm_xmap} class object returned by \code{\link[=read_xmap]{read_xmap()}}.}

\item{qnt}{\code{qm_qnt} class object returned by \code{\link[=read_qnt]{read_qnt()}}.}

\item{cluster}{\code{qm_cluster} class object returned by \code{\link[=cluster_xmap]{cluster_xmap()}}.}

\item{maps_x, maps_y}{Sizes of maps along x- and y-axes comprising guide net map.
(default: \code{NULL}).}

\item{fine_phase}{A character vector to specify fine grained phases
which tend to comprise multi-phase pixels e.g., c('Pl', 'Amp')}

\item{fine_th}{A threshold of membership degrees to 0.9}

\item{fixAB}{Fix AB in case compositions of a mineral is constant (default: \code{NULL}).}

\item{fixB}{Fix B (default: \code{NULL}).}

\item{saving}{\code{TRUE} (default) saves the results into \code{qntmap} directory under
the directory \code{xmap} is read from. \code{FALSE} does not save.`}
}
\description{
Quantify X-ray maps
}
