% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_cnd.0_cnd}
\alias{read_cnd.0_cnd}
\title{method for read_cnd}
\usage{
\method{read_cnd}{0_cnd}(x, pattern = NULL, n = NULL)
}
\arguments{
\item{x}{input}

\item{pattern}{character vector. Used to extract rows which contains phrase matching pattern.}

\item{n}{integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.}
}
\description{
method for read_cnd
}
