% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_cnd.default}
\alias{read_cnd.default}
\title{default method for read_cnd}
\usage{
\method{read_cnd}{default}(x, ...)
}
\arguments{
\item{x}{path to the file}

\item{...}{other arguments passed to methods}
}
\description{
default method for read_cnd
}
