% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_cnd.map_cnd}
\alias{read_cnd.map_cnd}
\title{method for read_cnd}
\usage{
\method{read_cnd}{map_cnd}(x, ...)
}
\arguments{
\item{x}{returned value from readLines. each elements starts with "$"}

\item{...}{other arguments passed to methods}
}
\description{
method for read_cnd
}
