% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map_beam.R
\name{read_map_beam.0_cnd}
\alias{read_map_beam.0_cnd}
\title{read mapping beam conditions}
\usage{
\method{read_map_beam}{0_cnd}(x, pattern = c(dwell =
  "Dwell Time \\\\[msec\\\\]", beam_map =
  "Probe Current (Avg, Before After )?\\\\[A\\\\]"), n = c(39, 17))
}
\arguments{
\item{x}{x}

\item{pattern}{patterns to be matched}

\item{n}{nth lines}
}
\description{
read mapping beam conditions
}
