% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_map_beam}
\alias{read_map_beam}
\title{read mapping beam information from 0.cnd}
\usage{
read_map_beam(path = "0.cnd", pattern = c(dwell =
  "Dwell Time \\\\[msec\\\\]", beam_map =
  "Probe Current (Avg, Before After )?\\\\[A\\\\]"), n = c(39, 17))
}
\arguments{
\item{path}{a path to the .cnd file}

\item{pattern}{character vector. Used to extract rows which contains phrase matching pattern.}

\item{n}{integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.}
}
\description{
read mapping beam information from 0.cnd
}
