% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map_pos.R
\name{read_map_pos.0_cnd}
\alias{read_map_pos.0_cnd}
\title{read mapping stage information from 0.cnd}
\usage{
\method{read_map_pos}{0_cnd}(x,
  pattern = c("Measurement Start Position X",
  "Measurement Start Position Y", "X(-axis)? Step Number",
  "Y(-axis)? Step Number", "X(-axis)? Step Size", "Y(-axis)? Step Size"),
  n = c(27, 28, 30:33))
}
\arguments{
\item{x}{input}

\item{pattern}{character vector. Used to extract rows which contains phrase matching pattern.}

\item{n}{integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.}
}
\description{
read mapping stage information from 0.cnd
}
