% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map_pos.R
\name{read_map_pos.map_cnd}
\alias{read_map_pos.map_cnd}
\title{a method for read_map_pos}
\usage{
\method{read_map_pos}{map_cnd}(x, ...)
}
\arguments{
\item{x}{input}

\item{...}{other arguments passed to methods}
}
\description{
a method for read_map_pos
}
