% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-qnt-.R
\name{read_qnt}
\alias{read_qnt}
\title{compile quantitative data}
\usage{
read_qnt(
  wd = dir(pattern = ("(.*_QNT|^\\\\.qnt)$"), all.files = TRUE)[1],
  phase_list = NULL,
  conditions = NULL,
  saving = TRUE
)
}
\arguments{
\item{wd}{directory path containing .qnt files}

\item{phase_list}{A path to a csv file containing columns indicating
phase of each analysis and \code{TRUE} or \code{FALSE} to use it for quantifying.}

\item{conditions}{A path to a csv file which records analytical conditions of elements.
The csv file must have following columns while others are discarded:
\code{Oxide}, \code{Element}, \verb{Bg+ [mm]}, \verb{Bg- [mm]}, \code{Peak [sec]}, and \code{Bg [sec]}.
\code{NULL} (default) retrieves analytical conditions from
\code{elem.qnt}, \code{elint.qnt}, and \code{.cnd/elemw.cnd} / \code{Pos_0001/data001.cnd}.}

\item{saving}{\code{TRUE} (defau;t) or \code{FALSE} to save phase list.}
}
\description{
compile quantitative data
}
