% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qnt_elemw.R
\name{read_qnt_elemw.0_cnd}
\alias{read_qnt_elemw.0_cnd}
\title{read all .cnd files under the specified directory}
\usage{
\method{read_qnt_elemw}{0_cnd}(x, pattern = c(bgp_pos =
  "(Back |BG)\\\\+\\\\[mm\\\\]", bgm_pos = "(Back |BG)-\\\\[mm\\\\]", pk_t
  = "(Peak|Pk)( Meas\\\\.)? (Time|t)", bg_t =
  "(Back|BG)( Meas\\\\.)? (Time|t)"), n = c(14, 15, 17, 18), each = 21)
}
\arguments{
\item{x}{path to the file}

\item{each}{n appears in each rows}
}
\description{
read all .cnd files under the specified directory
}
