% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_qnt_elemw}
\alias{read_qnt_elemw}
\title{read all .cnd files under the specified directory}
\usage{
read_qnt_elemw(path = "./.qnt/.cnd/elemw.cnd", pattern = c(bgp_pos =
  "(Back |BG)\\\\+\\\\[mm\\\\]", bgm_pos = "(Back |BG)-\\\\[mm\\\\]", pk_t =
  "(Peak|Pk)( Meas\\\\.)? (Time|t)", bg_t = "(Back|BG)( Meas\\\\.)? (Time|t)"),
  n = c(14, 15, 17, 18), each = 21)
}
\arguments{
\item{path}{a path to the .cnd file}

\item{pattern}{character vector. Used to extract rows which contains phrase matching pattern.}

\item{n}{integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.}

\item{each}{n appears in each rows}
}
\description{
read all .cnd files under the specified directory
}
