% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-xmap-.R
\name{read_xmap}
\alias{read_xmap}
\title{Read X-ray map data}
\usage{
read_xmap(wd = ".map/1", DT = 0, renew = FALSE, saving = TRUE,
  .map = "(data[0-9]+\\\\.csv)|([1-9][0-9]*_map\\\\.txt)",
  .cnd = "(data[0-9]+|[1-9][0-9]*)\\\\.cnd")
}
\arguments{
\item{wd}{directory path containing mapping data (e.g., ./.map/.1)}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{renew}{\code{TRUE} (default) read all data from original files.
\code{FALSE} tries to read \code{xmap.RDS} exists in \code{wd}, instead.}

\item{saving}{\code{TRUE} (default) or \code{FALSE} to save the result as \code{xmap.RDS} file in \code{wd}.}

\item{.map, .cnd}{Regular expressions to match file names of
ASCII converted maps (\code{.map}) and condition files (\code{.cnd})}
}
\description{
Read X-ray map data
}
