% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove-outlier.R
\name{remove_outlier}
\alias{remove_outlier}
\title{Remove outliers caused by multi-phase pixels}
\usage{
remove_outlier(epma, interval = c("prediction", "tukey"),
  method = c("rq", "lsft", "median"), percentile = 0.99)
}
\arguments{
\item{epma}{A value returned by \code{tidy_epma()}}

\item{interval}{A type of the interval. Data points outside intervals are treated as outliers.
If \code{"prediction"`` (default), prediction intervals are used based on Poisson process. If }"tukey"``, conditional lower and upper whiskers are used where the lower is
$Q_{1/4} - 1.5 IQR$ and the upper is $Q_{3/4} + 1.5 IQR$.}

\item{method}{Applicable when \code{interval = "prediction"}.
If \code{"rq"`` (default), quantile regression is performed. If }"lsfit"`, least square regression is performed.}

\item{percentile}{A percentile of predictive interaval.
Applicable when \code{interval = "prediction"}.}
}
\description{
When comparing data points of spot analysis and mapping analysis who share
the same coordinates, it is expected that they analyze the same phases.
However, this is unlikely when mapping analysis involves multi-phase pixels
due to sizes of pixels and phases. This function removes such multi-phase
pixels by finding outliers of mapping peak intensities.
}
