% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save4qm}
\alias{save4qm}
\alias{save4qm.default}
\alias{save4qm.data.frame}
\alias{save4qm.qm_cluster}
\alias{save4qm.qntmap}
\title{S3 generics to save returned values from functions in QntMap package.}
\usage{
save4qm(x, nm = "", saving = TRUE, ...)

\method{save4qm}{default}(x, nm, saving, ...)

\method{save4qm}{data.frame}(x, nm, saving, ...)

\method{save4qm}{qm_cluster}(x, nm, saving, ...)

\method{save4qm}{qntmap}(x, nm, saving, ...)
}
\arguments{
\item{x}{object to save}

\item{nm}{name of object to save}

\item{saving}{\code{TRUE} or \code{FALSE} to save result file(s).}

\item{...}{other arguments passed to class methods}

\item{dir_qntmap}{A path to the directory to save \code{x}.}
}
\value{
\code{invisible(x)}
}
\description{
S3 generics to save returned values from functions in QntMap package.
}
\section{\code{save4qm.default}}{

A default method.
This is a wrapper of \code{saveRDS} together with invisible return of \code{x}.
}

\section{\code{save4qm.data.frame}}{

A method for \code{data.frame} class.
This is a wrapper of \code{data.table::fwrite}
together with invisible return of \code{x}.
}

\section{\code{save4qm.qm_cluster}}{

A method for \code{qm_cluster} class returned by
\code{\link[=cluster_xmap]{cluster_xmap()}} or \code{\link[=group_cluster]{group_cluster()}}.
Object is saved as RDS file and png file.
The latter shows distribution of phases among a mapped area.
}

\section{\code{save4qm.qntmap}}{

A method for \code{qntmap} class returned by \code{qntmap()} or \code{quantify()}
}

