% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment.R
\name{segment}
\alias{segment}
\alias{segment.character}
\title{Segment mapping area}
\usage{
segment(x, ...)

\method{segment}{character}(x, ...)
}
\arguments{
\item{x}{A path to the PNG image file.}

\item{...}{ignored}
}
\value{
A character vector indicating pixel colors in RGBA style for input image.
}
\description{
Make index to segment mapping area.
The obtained index is further utilized in \code{\link[=mean.qntmap]{mean.qntmap()}}.
}
\examples{
x <- system.file("img", "Rlogo.png", package="png")
head(segment(x))

}
\seealso{
\code{\link[=mean.qntmap]{mean.qntmap()}}
}
