% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_qltmap_cls.R
\name{summary.qltmap_cls}
\alias{summary.qltmap_cls}
\title{summary qltmap_cls data}
\usage{
\method{summary}{qltmap_cls}(object, ...)
}
\arguments{
\item{object}{qntmap class data}

\item{...}{other arguments are discarded}
}
\description{
Returns abundance ratios of clusters.
}
