% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_qntmap.R
\name{summary.qntmap}
\alias{summary.qntmap}
\title{summary qntmap class data.}
\usage{
\method{summary}{qntmap}(object, ...)
}
\arguments{
\item{object}{qntmap class data}

\item{...}{other arguments are discarded}
}
\description{
summary qntmap class data.
}
