% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary_methods}
\alias{summary_methods}
\alias{summary.qm_cluster}
\title{summary methods}
\usage{
\method{summary}{qm_cluster}(object, ...)
}
\arguments{
\item{object}{object to be summarized}

\item{...}{other arguments to control summary results}
}
\description{
summary methods
}
\section{summary.qm_cluster}{
 Returns abundance ratios of clusters.
}

