% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy-epma.R
\name{tidy_epma}
\alias{tidy_epma}
\title{Compile epma data into a data frame}
\usage{
tidy_epma(qnt, xmap, cluster = NULL, subcluster = TRUE,
  suffix = "_.*")
}
\arguments{
\item{qnt}{object returned by read_qnt}

\item{xmap}{object returned by read_xmap}

\item{cluster}{object returned by cluster_xmap}

\item{subcluster}{Calculate parameters for
sub-clusters (TRUE) or for super-clusters (FALSE).}

\item{suffix}{A regular expression of suffix to identify sub-clusters.}
}
\description{
Compile epma data into a data frame
}
