% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster.R
\name{cluster_xmap}
\alias{cluster_xmap}
\title{Cluster mapping data into mineral species}
\usage{
cluster_xmap(xmap, centers, elements = intersect(names(xmap),
  colnames(centers)), saving = TRUE, group_cluster = FALSE)
}
\arguments{
\item{xmap}{A \code{qm_xmap} class object returned by \code{\link[=read_xmap]{read_xmap()}}}

\item{centers}{c-by-p matrix returned by \code{\link[=find_centers]{find_centers()}} or by manually;
c clusters and p features.
Used to guess initial centers (or centroids) of clusters.
A value returned by , typically \code{\link{data.frame}} or \code{\link{matrix}},
indicating initial guess centers (or centroids) or clusters.
See \code{\link[=find_centers]{find_centers()}}.}

\item{elements}{A character vector to chose elements to be utilized in cluster analysis.
\code{NULL} (default) selects as much elements as possible.}

\item{saving}{\code{TRUE} or \code{FALSE} to save result.}

\item{group_cluster}{\code{FALSE} (default) or \code{TRUE} to integrate same phase subgrouped using suffix.
For example,
clusters named "Pl_NaRich" and "Pl_NaPoor" are integrated to "Pl" cluster .}
}
\description{
Cluster mapping data into mineral species
}
