% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epma_tidy.R
\name{epma_tidy}
\alias{epma_tidy}
\title{compile epma data}
\usage{
epma_tidy(wd = NULL, dir_map = NULL, RDS_cluster = NULL,
  qnt = qnt_load(wd), qltmap = if (is.null(dir_map)) NULL else
  qltmap_load(dir_map), cluster = if (is.null(RDS_cluster)) NA else
  readRDS(RDS_cluster))
}
\arguments{
\item{wd}{path to the directory containing .qnt directory}

\item{dir_map}{path to the directory containing mapping analysis e.g., ./.map/1/}

\item{RDS_cluster}{path to the clustering result}

\item{qnt}{object returned by qnt_load}

\item{qltmap}{object returned by qltmap_load}

\item{cluster}{object returned by qltmap_cls_pois}
}
\description{
compile epma data
}
