% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-centers-.R
\name{find_centers}
\alias{find_centers}
\title{Initialize centroids for \code{\link[=cluster_xmap]{cluster_xmap()}}}
\usage{
find_centers(xmap, qnt, fine_phase = NULL, saveas = "centers0.csv",
  epma = tidy_epma(qnt, xmap), ...)
}
\arguments{
\item{xmap}{An object generated by \code{\link[=read_xmap]{read_xmap()}}}

\item{qnt}{An object generated by \code{\link[=read_qnt]{read_qnt()}}`}

\item{fine_phase}{A character vector specifying fine grained phases who should comprise
multi-phase pixels.}

\item{saveas}{File name to save result. \code{FALSE} if not saving.}

\item{...}{Arguments passed on to \code{remove_outlier}
\describe{
  \item{interval}{A type of the interval. Data points outside intervals are treated as outliers.
If \code{"prediction"`` (default), prediction intervals are used based on Poisson process. If }"tukey"``, conditional lower and upper whiskers are used where the lower is
$Q_{1/4} - 1.5 IQR$ and the upper is $Q_{3/4} + 1.5 IQR$.}
  \item{method}{Applicable when \code{interval = "prediction"}.
If \code{"rq"} (default), quantile regression is performed (\code{quantreg::rq}).
If \code{"lsfit"}, least square regression is performed (\code{stats::lsfit}).}
  \item{percentile}{A percentile of predictive interaval.
Applicable when \code{interval = "prediction"}.}
}}
}
\description{
Initial centroids are guessed by comparing data points from the quantified
spots and mapped pixels from the same coordinates.
First, imputation is performed for mapping intensities counted from
multi-phase pixels (See "Multi-phase pixels" section).
Second, for each phase, \code{\link{median}} values of X-ray maps are utilized as
initial centroids.
Third, if there are missing values especially for intensities of electrons
(secondary, backscatter, ...), imputation is performed for them
(See "Impute missings" section).
}
\section{Multi-phase pixels}{

X-ray mapping inevitably involves multi-phase pixels, especially for
fine-grained phases. When the multi-phase pixels are member of training
data set, guessing initial centroids can be biased. This is why guessing
centroids is based on \code{\link{median}} instead of \code{\link{mean}}. However, if pixels
to guess centroids for cetrtain phase are comprising a large number of or
even a full of multi-phase pixels, \code{\link{median}} is still vulnerable. For
example, refer to a reprinted figure 5 from
\href{https://doi.org/10.2138/am-2018-6323CCBY}{Yasumoto et al. (2018)}, and
captions in original article. Phases such as Amp, Di, and Pl are outlying
the regression curves due to multi-phase pixels. For these phases,
median values of peak X-ray intensities from X-ray mapping are unreliable
values for initial centroids. Thus, data points from multi-phase pixels
should be regarded as outliers, and imputetions should be performed on them.
In qntmap, mapping intensities are simply substituted by spot intensities.

Outliers are identified by examining if confidence intervals data points
overlap with predictive intervals or conditional whiskers of Tukey's choice.
}

\section{Impute missings}{

When spot analysis on some phases are performed outside the mapping area,
X-ray intensities under mapping conditions can be calculated from
quantified peak intensities. However, intensities of secondary and backscatter
electrons cannot be calculated because electrons are generally not
analyzed in spot analysis. Thus, imputation is performed. First, calculate
median values of mapping X-ray intensities. Second, find a pixel in the map
which have closest value to the median values. Third, retrive electron
intensities from that pixel and utilize as a centroid.

Note that missing values may remain if certain elements are analyzed in
spot analysis but not in mapping analysis.
}

\references{
Yasumoto, A., Yoshida, K., Kuwatani, T., Nakamura, D., Svojtka, M., &
Hirajima, T. (2018).
A rapid and precise quantitative electron probe chemical mapping technique
and its application to an ultrahigh-pressure eclogite
from the Moldanubian Zone of the Bohemian Massif
(Nové Dvory, Czech Republic).
American Mineralogist, 103(10), 1690-1698,
\url{https://doi.org/10.2138/am-2018-6323CCBY}.
}
