% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean-angled.R
\name{hmean}
\alias{hmean}
\alias{vmean}
\title{Calculate means for horizontal and vertical directions}
\usage{
hmean(x, cluster = NULL, density = NULL, ..., .by = "y")

vmean(x, cluster = NULL, density = NULL, ..., .by = "x")
}
\arguments{
\item{x}{A \code{qntmap} class object returned by \code{\link[=quantify]{quantify()}} or \code{\link[=qntmap]{qntmap()}}.}

\item{cluster}{A result of \code{cluster_xmap()} (i.e., \code{qm_cluster} class object).
\code{cluster} is required when specifying \code{density}.}

\item{density}{A named numeric vector specifying density of all the phases in \code{x}.
(e.g., \code{c(Qtz = 2.6, Ab = 2.6)}).}

\item{...}{Ignored}

\item{.by}{Internal use. Do not change.
"y" for \code{hmean} to calculate mean from pixels with same \code{y}-coordinates,
and "x" for \code{vmean} to calculate mean from pixels with same \code{x}-coordinates.}
}
\description{
Calculate means for horizontal and vertical directions for mapping data.
}
\seealso{
mean.qntmap
}
