% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\alias{mean.qntmap}
\title{Arithmetric mean for \code{qntmap} package}
\usage{
\method{mean}{qntmap}(x, index = "Whole area", ...)
}
\arguments{
\item{x}{A \code{qntmap} class object returned by \code{\link[=quantify]{quantify()}} or \code{\link[=qntmap]{qntmap()}}.}

\item{index}{a vector of length `1`` or length equal to number of pixels in map.
An index can be created using mask image through \code{\link[=segment]{segment()}}.}

\item{...}{Other arguments passed to \code{\link[base:mean.default]{base::mean.default()}}}
}
\description{
S3 methods for the arithmetric mean.
See \code{\link[base:mean]{base::mean()}}  for general use of \code{plot}.
}
\section{mean.qntmap}{

returns a \code{data.frame} whose first column lists elements in \code{qntmap} class object.
Second to last columns lists \code{mean} values of each elements for each index.
}

\examples{
# Calculate mean for qntmap class object

## mapping data corresponds of 100 x 2 pixels
## element A in the left half of the map yields 
## average 50 wt\% with standard deviation = 1.
## and in the right half of the map yields 
## average 50 wt\% with standard deviation = 1.
qm <- list(
  A = list(wt = data.frame(rnorm(100, 50), rnorm(100, 20)))
)
class(qm) <- c('qntmap', 'list')

## If index is not specified, 
## mean value is calculated from whole area,
## which in this case gives a value around 35.
mean(qm)

## If index is given, 
## mean is calculated from the pixels sharing the same index.
## For example, pixels in the left-half of the map is labelled by 'L',
## and the other side by 'R'.
## In such a case, mean of A is around 50 fo 'L' and 30 for 'R'
mean(qm, index = rep(c('L', 'R'), each = 100))

}
\seealso{
\code{\link[base:mean]{base::mean()}}

\code{\link[=segment]{segment()}}

\code{\link[=quantify]{quantify()}}

\code{\link[=qntmap]{qntmap()}}
}
