% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_qntmap.R
\name{print.qntmap}
\alias{print.qntmap}
\title{print qntmap class data.}
\usage{
\method{print}{qntmap}(x, summarizing = TRUE, ...)
}
\arguments{
\item{x}{qntmap class data}

\item{summarizing}{TRUE in default will visibly return summary of x and invisibly return x. FALSE will just return x.}

\item{...}{other arguments are discarded}
}
\description{
print qntmap class data.
}
