% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.qntmap}
\alias{print.qntmap}
\title{\code{print} method for \code{qntmap} class object}
\usage{
\method{print}{qntmap}(x, summarizing = TRUE, ...)
}
\arguments{
\item{x}{\code{qntmap} class object returned by \code{quantify()} or \code{qntmap()}.}

\item{summarizing}{\code{TRUE} or \code{FALSE} to summarize x (default: \code{TRUE}.}

\item{...}{Discarded.}
}
\description{
\code{print} method for \code{qntmap} class object
}
