% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.qntmap}
\alias{print.qntmap}
\title{\code{print} method for \code{qntmap} class object}
\usage{
\method{print}{qntmap}(x, ...)
}
\arguments{
\item{x}{A \code{qntmap} class object returned by \code{\link[=quantify]{quantify()}} or \code{\link[=qntmap]{qntmap()}}.}

\item{...}{Discarded.}
}
\description{
\code{print} method for \code{qntmap} class object
}
