% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_centers.R
\name{qltmap_cls_centers}
\alias{qltmap_cls_centers}
\title{Generate initial centroids for clustering}
\usage{
qltmap_cls_centers(path_qnt = "qnt.csv", path_cnd = "0.cnd",
  qltmap = NULL, wd = NULL)
}
\arguments{
\item{path_qnt}{path to the csv file of quantitative analysis}

\item{path_cnd}{path to the 0.cnd file  of mapping analysis}

\item{qltmap}{path to the .RDS file which compiles mapping data}

\item{wd}{working directory which contains mapping data}
}
\description{
Generate initial centroids for clustering
}
