% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_centers.R
\name{qltmap_cls_centers}
\alias{qltmap_cls_centers}
\title{(DEPRECATED) Use find_centers}
\usage{
qltmap_cls_centers(wd = NULL, dir_map, phase_fine = NULL,
  qnt = read_qnt(paste0(wd, "/.qnt")), qltmap = read_xmap(dir_map),
  saving = "centers_initial0.csv")
}
\arguments{
\item{wd}{working directory which contains .qnt directory. Use current directory if NULL}

\item{dir_map}{path to the directory which contains mapping data.}

\item{phase_fine}{fine grained phases which tend to be appear in multi-phase pixels e.g., c('Pl', 'Amp')}

\item{qnt}{an object generated by qnt_load or a path to the .RDS file created by qnt_load}

\item{qltmap}{an object generated by qltmap_load or a path to the .RDS file created by qltmap_load}

\item{saving}{file name to save. FALSE if not saving.}
}
\description{
(DEPRECATED) Use find_centers
}
