% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_centers.R
\name{qltmap_cls_centers}
\alias{qltmap_cls_centers}
\title{Generate initial centroids for clustering}
\usage{
qltmap_cls_centers(qnt = NULL, qltmap = NULL, wd = NULL, dir_map = NULL,
  cnd_qltmap = paste0(dir_map, "/0.cnd"), saving = "centers_initial0.csv")
}
\arguments{
\item{qnt}{an object generated by qnt_load or a path to the .RDS file created by qnt_load}

\item{qltmap}{an object generated by qltmap_load or a path to the .RDS file created by qltmap_load}

\item{wd}{working directory which contains .qnt directory. Use current directory if NULL}

\item{dir_map}{path to the directory which contains mapping data.}

\item{cnd_qltmap}{path to the condition file of mapping data}

\item{saving}{file name to save. FALSE if not saving.}
}
\description{
Generate initial centroids for clustering
}
