% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_cls_pois.R
\name{qltmap_cls_pois}
\alias{qltmap_cls_pois}
\title{cluster mapping data into mineral species}
\usage{
qltmap_cls_pois(centers_initial, qltmap = NULL, elements = NULL,
  wd = NULL, saving = TRUE, integration = TRUE)
}
\arguments{
\item{centers_initial}{a path to csv file telling initial centers of clusters}

\item{qltmap}{default to NULL}

\item{elements}{If NULL, all mapped elements are used for clustering. Specifiyng elements may reduce analytical time.}

\item{wd}{Directory containing mapping data. If NULL, current directory is wd.}

\item{saving}{TRUE or FALSE to save result}

\item{integration}{TRUE or FALSE to integrate same phase with obiously different compositions. For example, when there are clusters named as Pl_NaRich and Pl_NaPoor, they are integrated to Pl.}
}
\description{
cluster mapping data into mineral species
}
