% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmap.R
\name{qltmap_load}
\alias{qltmap_load}
\title{DEPRECATED!! use read_xmap}
\usage{
qltmap_load(wd = ".", DT = 0, RDS = "xmap.RDS", renew = FALSE,
  saving = TRUE, .map = "(data)?[0-9]*[1-9](\\\\.csv|_map\\\\.txt)",
  .cnd = "(data)?[0-9]*[1-9]\\\\.cnd")
}
\arguments{
\item{wd}{working directory which contains mapping data}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{RDS}{name of RDS file to be saved/readed}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}
}
\description{
DEPRECATED!! use read_xmap
}
