% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qltmap_load.R
\name{qltmap_load}
\alias{qltmap_load}
\title{load mapping data}
\usage{
qltmap_load(wd = NULL, RDS = "qltmap.RDS", list_of = c("vector", "matrix",
  "column"), renew = FALSE, saving = TRUE)
}
\arguments{
\item{wd}{working directory which contains mapping data}

\item{RDS}{name of RDS file to be saved/readed}

\item{list_of}{structure of output data}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}
}
\description{
load mapping data
}
