% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qnt.R
\name{qnt_load}
\alias{qnt_load}
\title{DEPRECATED!! Use read_qnt}
\usage{
qnt_load(wd = NULL, RDS = "qnt.RDS", phase_list = NULL,
  renew = FALSE, saving = TRUE)
}
\arguments{
\item{wd}{path to the directory containing .qnt directory}

\item{RDS}{name of RDS file to be saved/readed}

\item{phase_list}{path to the csv file containing columns indicating phase of each analysis and true or false to use it for quantifying.}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}
}
\description{
DEPRECATED!! Use read_qnt
}
