% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnt_load.R
\name{qnt_load}
\alias{qnt_load}
\title{compile quantitative data}
\usage{
qnt_load(wd = NULL)
}
\arguments{
\item{wd}{path to the directory containing .qnt directory}
}
\description{
compile quantitative data
}
