% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap.R
\name{qntmap}
\alias{qntmap}
\title{Interactively quantify X-ray maps}
\usage{
qntmap(
  xmap_dir = NULL,
  qnt_dir = NULL,
  deadtime = 1100,
  phase_list = NULL,
  console = FALSE,
  selected = c("auto", "Input", "Map", "Spot")
)
}
\arguments{
\item{xmap_dir, qnt_dir}{Paths to the directories containing data from mapping and spot analysis.}

\item{deadtime}{Dead time in nano seconds.}

\item{phase_list}{A path to a csv file containing columns indicating
phase of each analysis and \code{TRUE} or \code{FALSE} to use it for quantifying.}

\item{console}{Wheter to use console-based interactive mode (default: \code{FALSE}).
Notet hat the console-based one is retiring and no more maintained.}

\item{selected}{A starting page of shiny UI. One of the "auto", "Input", "Map" or "Spot".
If auto (default), "Map" is chosen when \code{xmap_dir} is specified, or "Spot"
when \code{qnt_dir} is specified, or "Input" when both are not specified.}
}
\description{
By default, it starts browser to open web-based UI. All arguments are optional.
}
