% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap.R
\name{qntmap}
\alias{qntmap}
\title{Interactively quantify X-ray maps}
\usage{
qntmap(xmap_dir = NULL, qnt_dir = NULL, deadtime = 1100,
  phase_list = NULL, shiny = TRUE)
}
\description{
Interactively quantify X-ray maps
}
