% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap_quantify.R
\name{qntmap_quantify}
\alias{qntmap_quantify}
\title{quantify qualtitative mapping data}
\usage{
qntmap_quantify(wd = NULL, dir_map = NULL, maps_x = NULL, maps_y = NULL,
  path_phase = "qnt_phase.csv", RDS_cluster = NULL, fine_phase = NULL,
  fine_th = 0.9)
}
\arguments{
\item{wd}{working directory which contains .qnt and .map directories}

\item{dir_map}{directory containing map data to be quantified}

\item{maps_x}{aaa}

\item{maps_y}{aaa}

\item{path_phase}{aaa}

\item{RDS_cluster}{path to the output RDS file of clustering. NULL in default look for the newest one in dir_map/clustering}

\item{fine_phase}{aaa}

\item{fine_th}{0.9}
}
\description{
quantify qualtitative mapping data
}
