% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{qntmap_quantify}
\alias{qntmap_quantify}
\title{(Deprecated) Use quantify.}
\usage{
qntmap_quantify(wd = ".", dir_map, RDS_cluster, maps_x = NULL,
  maps_y = NULL, fine_phase = NULL, fine_th = 0.9,
  qnt = qnt_load(wd), qltmap = qltmap_load(dir_map),
  cluster = readRDS(RDS_cluster))
}
\arguments{
\item{wd}{working directory which contains .qnt and .map directories}

\item{maps_x}{A x-axis size of maps comprising guide net map (default: \code{NULL}).}

\item{maps_y}{A y-axis size of maps comprising guide net map (default: \code{NULL}).}

\item{fine_phase}{A character vector to specify fine grained phases which tend to comprise multi-phase pixels e.g., c('Pl', 'Amp')}

\item{fine_th}{0.9}

\item{qnt}{\code{qm_qnt} class object returned by \code{read_qnt()}.}

\item{cluster}{\code{qm_cluster} class object returned by \code{cluter_xmap()}.}
}
\description{
(Deprecated) Use quantify.
}
