% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qntmap_quantify.R
\name{qntmap_quantify}
\alias{qntmap_quantify}
\title{quantify qualtitative mapping data}
\usage{
qntmap_quantify(wd = NULL, dir_map, RDS_cluster, maps_x = NULL,
  maps_y = NULL, fine_phase = NULL, fine_th = 0.9, qnt = qnt_load(wd),
  qltmap = qltmap_load(dir_map), cluster = readRDS(RDS_cluster))
}
\arguments{
\item{wd}{working directory which contains .qnt and .map directories}

\item{dir_map}{directory containing map data to be quantified}

\item{RDS_cluster}{path to the output RDS file of clustering. NULL in default look for the newest one in dir_map/clustering}

\item{maps_x}{x of maps. Assign when you use guide net map.}

\item{maps_y}{y of maps. Assign when you use guide net map.}

\item{fine_phase}{fine-grained phases which tend to be appear in multi-phase pixels}

\item{fine_th}{0.9}

\item{qnt}{object of class qnt}

\item{qltmap}{object of class qltmap}

\item{cluster}{object of class PoiClaClu}
}
\description{
quantify qualtitative mapping data
}
