% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantify X-ray maps}
\usage{
quantify(xmap, qnt, cluster, subcluster = TRUE, suffix = "_.*",
  maps_x = attr(xmap, "pixel")[1L], maps_y = attr(xmap, "pixel")[2L],
  fine_phase = NULL, fine_th = 0.9, fix = NULL, se = FALSE,
  saving = TRUE)
}
\arguments{
\item{xmap}{\code{qm_xmap} class object returned by \code{\link[=read_xmap]{read_xmap()}}.}

\item{qnt}{\code{qm_qnt} class object returned by \code{\link[=read_qnt]{read_qnt()}}.}

\item{cluster}{\code{qm_cluster} class object returned by \code{\link[=cluster_xmap]{cluster_xmap()}}.}

\item{subcluster}{Calculate parameters for
sub-clusters (TRUE) or for super-clusters (FALSE).}

\item{suffix}{A regular expression of suffix to identify sub-clusters.}

\item{maps_x, maps_y}{Sizes of maps along x- and y-axes comprising guide net map.
(default: \code{NULL}).}

\item{fine_phase}{Deprecated as of qntmap > 0.4.0. Use \code{phase} instead.}

\item{fine_th}{A threshold of membership degrees to 0.9}

\item{fix}{A path to the file specifying chemical compositions of
some elements in some phases (default: \code{NULL}).}

\item{se}{\code{FALSE} in default, and is forced when \code{fix} is specified.
\code{TRUE} calculates standard errors, but require large memories.}

\item{saving}{\code{TRUE} (default) saves the results into \code{qntmap} directory under
the directory \code{xmap} is read from. \code{FALSE} does not save.`}
}
\description{
Quantify X-ray maps
}
