% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify.R
\name{quantify}
\alias{quantify}
\title{Quantify X-ray maps}
\usage{
quantify(
  xmap,
  qnt,
  cluster,
  maps_x = attr(xmap, "pixel")[1L],
  maps_y = attr(xmap, "pixel")[2L],
  phase = everything(),
  element = everything(),
  fine_phase = NULL,
  fine_th = 0.9,
  fix = NULL,
  se = FALSE,
  saving = TRUE,
  ...
)
}
\arguments{
\item{xmap}{Object returned by \code{\link[=read_xmap]{read_xmap()}}.}

\item{qnt}{Object returned by \code{\link[=read_qnt]{read_qnt()}}.}

\item{cluster}{Object returned by \code{\link[=cluster_xmap]{cluster_xmap()}}.}

\item{maps_x, maps_y}{Sizes of maps along x- and y-axes comprising guide net map.
(default: \code{NULL}).}

\item{phase}{Selected ones are referenced to detect outliers. Default selects everything.
Tidy selection is available. For example \code{c(Si, Ti)} selects them, and
\code{c(-Si, -Ti)} selects everything except them.}

\item{element}{Selected ones are referenced to detect outliers. Default selects everything.
Tidy selection is available. For example \code{c(Si, Ti)} selects them, and
\code{c(-Si, -Ti)} selects everything except them.}

\item{fine_phase}{Deprecated as of qntmap > 0.4.0. Use \code{phase} instead.}

\item{fine_th}{A threshold of membership degrees to 0.9}

\item{fix}{A path to the file specifying chemical compositions of
some elements in some phases (default: \code{NULL}).}

\item{se}{\code{FALSE} in default, and is forced when \code{fix} is specified.
\code{TRUE} calculates standard errors, but require large memories.}

\item{saving}{\code{TRUE} (default) saves the results into \code{qntmap} directory under
the directory \code{xmap} is read from. \code{FALSE} does not save.`}

\item{...}{
  Arguments passed on to \code{\link[=find_outlier]{find_outlier}}
  \describe{
    \item{\code{interval}}{A type of the interval. Data points outside intervals are treated as outliers.
If \code{"prediction"} (default), prediction intervals are used based on Poisson process.
If `"tukey"``, conditional lower and upper whiskers are used where the lower is
$Q_{1/4} - 1.5 IQR$ and the upper is $Q_{3/4} + 1.5 IQR$.}
    \item{\code{method}}{Applicable when \code{interval = "prediction"}.
If \code{"rq"} (default), quantile regression is performed (\code{\link[quantreg:rq]{quantreg::rq()}}).
If \code{"lsfit"}, least square regression is performed (\code{\link[stats:lsfit]{stats::lsfit()}}).}
    \item{\code{percentile}}{A percentile of predictive interaval.
Applicable when \code{interval = "prediction"}.}
  }}
}
\description{
Quantify X-ray maps
}
