% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cnd.R
\name{read_cnd}
\alias{read_cnd}
\title{read .cnd file of JEOL EPMA (0.cnd, 1.cnd, ...)}
\usage{
read_cnd(path = "0.cnd", pattern = ".", n = NULL)
}
\arguments{
\item{path}{a path to the .cnd file}

\item{pattern}{character vector. Used to extract rows which contains phrase matching pattern.}

\item{n}{integer vector of same length as pattern. Used to extract nth row of .cnd file in case pattern did not match any phrase.}
}
\description{
read .cnd file of JEOL EPMA (0.cnd, 1.cnd, ...)
}
