% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-cnd.R
\name{read_cnd}
\alias{read_cnd}
\alias{read_cnd.default}
\alias{read_cnd.map_cnd}
\alias{read_cnd.0_cnd}
\title{Read .cnd files}
\usage{
read_cnd(x, pattern = NULL, ...)

\method{read_cnd}{default}(x, pattern = NULL, ...)

\method{read_cnd}{map_cnd}(x, pattern = NULL, ...)

\method{read_cnd}{`0_cnd`}(x, pattern = NULL, n = NULL, ...)
}
\arguments{
\item{x}{A path to the file (e.g., '0.cnd', 'map.cnd', ...)}

\item{pattern}{pattern to substract}

\item{...}{Other arguments passed to methods}

\item{n}{integer vector of same length as pattern.
Used to extract nth row of .cnd file if no phrases match the \code{pattern}.}
}
\description{
Read .cnd files
}
\section{.default}{

A default method which returns a result of \code{\link[=readLines]{readLines()}}
with additional class according to the content of the file.
}

\section{read_cnd.map_cnd}{
 A method for \code{map_cnd} class object.
}

\section{read_cnd.0_cnd}{

A method for \verb{0_cnd} class object.
Used to extract rows which contains phrase matching pattern.
}

