% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-qnt-.R
\name{read_qnt}
\alias{read_qnt}
\title{compile quantitative data}
\usage{
read_qnt(wd = dir(pattern = ("(.*_QNT|^\\\\.qnt)$"), all.files =
  TRUE)[1], phase_list = NULL, renew = FALSE, saving = TRUE)
}
\arguments{
\item{wd}{directory path containing .qnt files}

\item{phase_list}{A path to a csv file containing columns indicating
phase of each analysis and \code{TRUE} or \code{FALSE} to use it for quantifying.}

\item{renew}{\code{TRUE} (default) read all data from original files.
\code{FALSE} tries to read \code{qnt.RDS} exists in \code{wd}, instead.}

\item{saving}{\code{TRUE} (default) or \code{FALSE} to save the result as \code{qnt.RDS} file in \code{wd}.}
}
\description{
compile quantitative data
}
