% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_xmap.R
\name{read_xmap}
\alias{read_xmap}
\title{read X-ray map data}
\usage{
read_xmap(wd = NULL, DT = 0, RDS = "qltmap.RDS", renew = FALSE,
  saving = TRUE)
}
\arguments{
\item{wd}{working directory which contains mapping data}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{RDS}{name of RDS file to be saved/readed}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}
}
\description{
read X-ray map data
}
