% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-xmap-.R
\name{read_xmap}
\alias{read_xmap}
\title{read X-ray map data}
\usage{
read_xmap(wd = ".map/1", DT = 0, renew = FALSE, saving = TRUE,
  .map = "(data[0-9]+\\\\.csv)|([1-9][0-9]*_map\\\\.txt)",
  .cnd = "(data[0-9]+|[1-9][0-9]*)\\\\.cnd")
}
\arguments{
\item{wd}{directory path containing mapping data (e.g., ./.map/.1)}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{renew}{if TRUE and the file specified by RDS exists, that file will be loaded}

\item{saving}{whether or not to save the data as RDS file}

\item{.map, .cnd}{regular expressions to match ASCII converted mapping results (\code{.map}) and condition files (\code{.cnd})}
}
\description{
read X-ray map data
}
