% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-xmap-.R
\name{read_xmap}
\alias{read_xmap}
\title{Read X-ray map data}
\usage{
read_xmap(wd = ".map/1", DT = 0, conditions = NULL,
  .map = "(data[0-9]+\\\\.csv)|([1-9][0-9]*_map\\\\.txt)",
  .cnd = "(data[0-9]+|[1-9][0-9]*)\\\\.cnd")
}
\arguments{
\item{wd}{directory path containing mapping data (e.g., ./.map/.1)}

\item{DT}{dead time in nano seconds (0 nsec in default)}

\item{conditions}{A path to a csv file which records file paths and analytical conditions of
mapping data.
Specifying \code{conditions} discards other parameters except \code{saving}.}

\item{.map, .cnd}{Regular expressions to match file names of
ASCII converted maps (\code{.map}) and condition files (\code{.cnd})}
}
\description{
Read X-ray map data
}
