% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save4qm}
\alias{save4qm}
\alias{save4qm.default}
\alias{save4qm.data.frame}
\alias{save4qm.qm_cluster}
\alias{save4qm.qntmap}
\title{Save objects created by qntmap package.}
\usage{
save4qm(x, nm = "", saving = TRUE, ...)

\method{save4qm}{default}(x, nm, saving, ...)

\method{save4qm}{data.frame}(x, nm, saving, ...)

\method{save4qm}{qm_cluster}(x, nm, saving, ...)

\method{save4qm}{qntmap}(x, nm, saving, ...)
}
\arguments{
\item{x}{An object to save.}

\item{nm}{A name of object to be saved as.}

\item{saving}{\code{TRUE} (default) or \code{FALSE} to save.}

\item{...}{Other arguments passed to class methods}
}
\value{
\code{invisible(x)}
}
\description{
Save objects created by qntmap package.
}
\section{\code{save4qm.default}}{

A default method saves an object as a binary RDS file.
}

\section{\code{save4qm.data.frame}}{

A data frame object is saved as a csv.
}

\section{\code{save4qm.qm_cluster}}{

A value returned by \code{\link[=cluster_xmap]{cluster_xmap()}} or \code{\link[=group_cluster]{group_cluster()}} is
saved as binary RDS file, dot-by-dot png file, and as svg file with a legend.
The png and svg files show distribution of phases among a mapped area.
}

\section{\code{save4qm.qntmap}}{

A value returned by \code{qntmap()} or \code{quantify()} is saved as
a binary RDS file, and csv files.
}

